+++ src/components/Icons/Icon.tsx
@@ -90,6 +90,7 @@ export type IconRotationTypes =
   | "rotate270";
 export type IconSizes =
   | "default"
+  | "xsmall"
   | "small"
   | "medium"
   | "large"